refactor(core): pipeline-worker persist 핸들러 분리
615줄 거대 핸들러 분해 (2/4) — persist 분기를 별도 파일로 추출.
sourceId→dbId 매핑, keyword linkage, whisper enqueue, embedding,
classify trigger 순서 보존. 외부 시그니처 동일.

리팩토링 마스터플랜 Phase 2 PR 2-F.
